The compiler now parses the function definitions first and then returns to the beginning of the stream, allowing for calls to precede functions

+++ src/symtable.c
@@ -87,7 +87,7 @@ void DestroyList(SymtableListNode *list)
     DestroyNode(list);
 }
 
-Symtable *InitSymtable(size_t size)
+Symtable *InitSymtable(unsigned long size)
 {
     Symtable *symtable;
     if ((symtable = calloc(1, sizeof(Symtable))) == NULL)
